Fixed bomb on start, cleaned up lexer tokens.
git-svn-id: https://svn.red-tux.net/trunk/ruby/zabcon@378 0b82185e-49c8-4601-bd6e-747293a2f232

+++ zabcon.rb
@@ -232,8 +232,8 @@ class ZabconApp
         puts "Unable to access configuration file: #{config_file}"
       end
       config={}
-    rescue String=>e
-      if e=="NoConfig"
+    rescue RuntimeError=>e
+      if e.message=="NoConfig"
         puts "Unable to find a default configuration file"
         config={}
       else
@@ -247,10 +247,10 @@ class ZabconApp
     server_keys=["server","username","password","proxy_server",
                 "proxy_port","proxy_user","proxy_password"]
 
-    ServerCredentials.instance["global"]=config.select_keys(server_keys).merge({"name"=>"global"})
-
-    config.delete_keys(server_keys)
-
+    if !config.empty?
+      ServerCredentials.instance["global"]=config.select_keys(server_keys).merge({"name"=>"global"})
+      config.delete_keys(server_keys)
+    end
 
     config.each_pair { |k,v|
       if k.match(/(.+)\[(.+)\]\[(.+)\]/)
